1279. Disable RENDER extension support in the ATI(misc) driver when mibank       and/or shadowfb is used (Marc La France). 1278. Remove cfb & friends from the ATI driver (Marc La France). 1277. Some additional ATI PCI IDs (Marc La France). 1276. Make compilation of drivers produce the same object whether or not       DPMS and DGA extensions are being built.  Partial changes to do the same       for RENDER (Marc La France). 1275. Ensure `make depend` goes through the same set of subdirectories as       `make all` does.  Allows MakefileAdditions to contain depend rule(s)       (Marc La France). 1274. Move $(EXTRA_DEFINES) to the end of ALLDEFINES.  Ditto for       $(CXXEXTRA_DEFINES) and CXXDEFINES.  Allows easier overrides at the       subdirectory level (Marc La France). 1273. Add a pre-clean phase to `make World` if xc/xmakefile already exists       (Marc La France).
+ update formatted docs
+ build & warning fixes

+++ extras/ogl-sample/main/gfx/lib/glu/libnurbs/interface/glcurveval.cc
@@ -31,12 +31,13 @@
 ** published by SGI, but has not been independently verified as being
 ** compliant with the OpenGL(R) version 1.2.1 Specification.
 */
+/* $XFree86$ */
 
 /*
  * glcurveval.c++
  *
- * $Date: 2001/01/15 21:48:13 $ $Revision: 1.1 $
- * $Header: /vol1/history/xf86/xc/extras/ogl-sample/main/gfx/lib/glu/libnurbs/interface/glcurveval.cc,v 1.1 2001/01/15 21:48:13 dawes Exp $
+ * $Date: 2001/01/21 21:19:07 $ $Revision: 1.2 $
+ * $Header: /vol1/history/xf86/xc/extras/ogl-sample/main/gfx/lib/glu/libnurbs/interface/glcurveval.cc,v 1.2 2001/01/21 21:19:07 tsi Exp $
  */
 
 /* Polynomial Evaluator Interface */
@@ -149,8 +150,8 @@ OpenGLCurveEvaluator::mapgrid1f(long nu, REAL u0, REAL u1)
 {
   if(output_triangles)
     {
-      global_grid_u0 = u0;
-      global_grid_u1 = u1;
+      global_grid_u0 = (int)u0;
+      global_grid_u1 = (int)u1;
       global_grid_nu = nu;
     }
   else
